+++ client.go
@@ -1,7 +1,0 @@
-package mediawiki
-
-import (
-	"github.com/hashicorp/go-retryablehttp"
-)
-
-var defaultClient = retryablehttp.NewClient()