+++ config.go
@@ -9,7 +9,14 @@ import (
 
 // ProcessDumpConfig is a configuration for high-level Process*Dump functions.
 //
-// UserAgent is required.
+// Client is required.
+//
+// Client should set User-Agent header with contact information, e.g.:
+//
+//     client := retryablehttp.NewClient()
+//     client.RequestLogHook = func(logger retryablehttp.Logger, req *http.Request, retry int) {
+//     	req.Header.Set("User-Agent", "My bot (user@example.com)")
+//     }
 type ProcessDumpConfig struct {
 	URL                    string
 	CacheDir               string
@@ -17,6 +24,5 @@ type ProcessDumpConfig struct {
 	DecompressionThreads   int
 	JSONDecodeThreads      int
 	ItemsProcessingThreads int
-	UserAgent              string
 	Progress               func(context.Context, x.Progress)
 }